backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0
redis==5.0.8
celery==5.4.0
httpx==0.27.2
python-dotenv==1.0.1
GitPython==3.1.43
sse-starlette==2.1.3
orjson==3.10.7